Remove livechat service
This pr includes:
1. Move livechat tests to ui/livechat
2. Remove livechat service and related configurations
3. Redirect livechat.ly.fish to jel.ly.fish

Change-type: major
