+++ .balena/balena.yml
@@ -5,6 +5,3 @@ build-variables:
     ui:
       - SERVER_HOST=http://api.ly.fish.local
       - SERVER_PORT=80
-    livechat:
-      - SERVER_HOST=http://api.ly.fish.local
-      - SERVER_PORT=80
