+++ .gitignore
@@ -105,7 +105,6 @@ docs/typedoc/*
 
 # Front app environment configuration files
 apps/ui/env-config.js
-apps/livechat/env-config.js
 
 # Decrypted secret files
 .balena/secrets/integration_typeform_signature_key
