+++ .resinci.yml
@@ -10,10 +10,6 @@ docker:
       docker_repo: balena/jellyfish-ui
       publish: true
     - path: .
-      dockerfile: apps/livechat/Dockerfile
-      docker_repo: balena/jellyfish-livechat
-      publish: true
-    - path: .
       dockerfile: Dockerfile.ci
       docker_repo: balena/jellyfish-sut
       publish: false
