--- apps/livechat/.eslintrc.yml
@@ -1,3 +1,0 @@
-env:
-  browser: true
-  node: false
