--- apps/livechat/DESCRIPTION.markdown
@@ -1,1 +1,0 @@
-This is the demo project for developing chat-widget.
