--- apps/livechat/conf/nginx.conf
@@ -1,46 +1,0 @@
-# Nginx configuration for the UI container. This config file is used in Dockerfile.ui
-
-server {
-	listen       80;
-	server_name  localhost;
-
-	# Enable gzip
-	gzip on;
-	gzip_min_length 1024;
-	gzip_types text/plain text/css text/xml text/javascript application/x-javascript application/javascript application/xml;
-	gzip_vary on;
-	gzip_disable "MSIE [1-6]\.";
-
-	# Remove as much server info as possible without recompiling
-	server_tokens off;
-
-	# redirect server error pages to the static page /50x.html
-	error_page   500 502 503 504  /50x.html;
-	location = /50x.html {
-		root   /usr/share/nginx/html;
-	}
-
-	location = /404.html {
-		root /usr/share/nginx/html;
-		internal;
-
-		add_header Cache-Control "no-store, no-cache, must-revalidate";
-	}
-
-	location ~\.(js|css)$ {
-		root /usr/share/nginx/html;
-
-		 expires 1y;
-		 add_header Cache-Control "public";
-		 access_log off;
-	}
-
-	location / {
-		root /usr/share/nginx/html;
-		# Redirect unknown paths to index.html, to allow for paths handled by the UI
-		# to work on initial page load
-		try_files $uri /index.html;
-
-		add_header Cache-Control "no-store, no-cache, must-revalidate";
-	}
-}
