--- apps/livechat/lib/index.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8" />
-        <title>Jellyfish</title>
-		<meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=0.25, maximum-scale=12, user-scalable=yes">
-		<link rel="shortcut icon" href="/favicon.ico" type="image/x-icon">
-		<link rel="icon" href="/favicon.ico" type="image/x-icon">
-		<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/all.css" integrity="sha384-50oBUHEmvpQ+1lW4y57PTFmhCaXp0ML5d60M1M7uH2+nqUivzIebhndOJK28anvf" crossorigin="anonymous">
-        <style>
-            html,
-            body,
-            #app {
-                height: 100%;
-                margin: 0;
-            }
-
-            * {
-                box-sizing: border-box;
-            }
-        </style>
-				<script src="/env-config.js"></script>
-    </head>
-
-    <body>
-        <div id="app"></div>
-    </body>
-</html>
