--- apps/livechat/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "jellyfish-livechat",
-  "version": "0.0.1",
-  "codename": "cephea",
-  "private": true,
-  "description": "The Jellyfish Livechat App",
-  "engines": {
-    "node": ">=16.0.0"
-  },
-  "scripts": {
-    "clean": "rimraf dist",
-    "build": "npm run clean && NODE_ENV=production webpack --config=./webpack.config.js",
-    "lint": "balena-lint lib && depcheck --ignore-bin-package --ignores=@balena/*,babel-loader,typedoc",
-    "lint:fix": "balena-lint --fix lib",
-    "doc": "typedoc lib/",
-    "prepack": "npm run build",
-    "env": "mkdir -p ./dist/livechat && NODE_ENV=development PUBLIC_DIR='./' ./conf/env.sh && cp ./env-config.js ./dist/livechat/",
-    "dev": "npm run env && NODE_ENV=development webpack serve --config=./webpack.config.js --color"
-  },
-  "author": "Balena.io. <hello@balena.io>",
-  "license": "AGPL-3.0",
-  "dependencies": {
-    "@balena/jellyfish-chat-widget": "^14.0.10",
-    "@balena/jellyfish-client-sdk": "^10.1.8",
-    "@balena/jellyfish-ui-components": "^15.0.3",
-    "@use-it/event-listener": "^0.1.7",
-    "css-loader": "^5.2.7",
-    "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^4.5.2",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-router-dom": "^5.3.0",
-    "rendition": "^25.13.1",
-    "style-loader": "^2.0.0",
-    "webpack": "^4.46.0",
-    "webpack-bundle-analyzer": "^4.5.0"
-  },
-  "devDependencies": {
-    "@balena/jellyfish-config": "^2.0.2",
-    "@balena/lint": "^6.2.0",
-    "babel-loader": "^8.2.3",
-    "depcheck": "^1.4.3",
-    "esbuild-loader": "^2.18.0",
-    "eslint": "^7.32.0",
-    "rimraf": "^3.0.2",
-    "typedoc": "^0.22.13",
-    "typescript": "^4.6.2",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4"
-  }
-}
