--- apps/livechat/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "@balena/jellyfish-config/config/tsconfig.json",
-  "compilerOptions": {
-    "target": "es2016",
-    "lib": [ "ES2019.Object", "DOM.Iterable" ],
-    "noUnusedParameters": false,
-    "noUnusedLocals": false,
-    "jsx": "react",
-    "declaration": false
-  },
-  "typedocOptions": {
-    "out": "../../docs/typedoc/livechat"
-  }
-}
