+++ docker-compose.test.yml
@@ -8,15 +8,12 @@ services:
       dockerfile: Dockerfile.ci
     depends_on:
       - api
-      - livechat
       - postgres
       - ui
       - balena-mdns-publisher
     environment:
       - LOGLEVEL=crit
       - NO_LOG_SUCCESS=1
-      - LIVECHAT_HOST=http://livechat
-      - LIVECHAT_PORT=80
       - SERVER_HOST=http://api
       - SERVER_PORT=80
       - UI_HOST=http://ui
