+++ docker-compose.yml
@@ -1,22 +1,6 @@
 version: '2.1'
 
 services:
-  livechat:
-    image: balena/jellyfish-livechat
-    build:
-      context: .
-      dockerfile: ./apps/livechat/Dockerfile
-    depends_on:
-      - api
-      - balena-mdns-publisher
-    networks:
-      - internal
-    environment:
-      - SENTRY_DSN_UI=0
-      - SERVER_HOST=http://api.ly.fish.local
-      - SERVER_PORT=80
-      - NGINX_PORT=80
-      - LIVECHAT_PORT=80
   postgres:
     image: balena/open-balena-db:4.1.0
     restart: always
