+++ haproxy/haproxy.cfg
@@ -17,11 +17,8 @@ option forwardfor
 bind *:80
 http-request add-header X-Forwarded-Proto http
 
-acl host_ui_backend hdr_dom(host) -i jel.ly.fish.local
+acl host_ui_backend hdr_dom(host) -i jel.ly.fish.local livechat.ly.fish.local
 use_backend ui_backend if host_ui_backend
-
-acl host_livechat_backend hdr_dom(host) -i livechat.ly.fish.local
-use_backend livechat_backend if host_livechat_backend
 
 acl host_api_backend hdr_dom(host) -i api.ly.fish.local
 use_backend api_backend if host_api_backend
@@ -52,12 +49,6 @@ server redis redis:6379 resolvers docker-bridge-resolver resolve-prefer ipv4 che
 backend postgres_backend
 mode tcp
 server postgres postgres:5432 resolvers docker-bridge-resolver resolve-prefer ipv4 check port 5432
-
-backend livechat_backend
-mode http
-option forwardfor
-balance roundrobin
-server livechat livechat:80 resolvers docker-bridge-resolver resolve-prefer ipv4 check port 80
 
 backend api_backend
 mode http
