+++ package.json
@@ -13,26 +13,24 @@
     "node": ">=16.0.0"
   },
   "scripts": {
-    "lint": "eslint --ext .js,.jsx scripts test && npm run lint:check && npm run lint:deps && npm run lint:livechat && npm run lint:server && npm run lint:ui",
+    "lint": "eslint --ext .js,.jsx scripts test && npm run lint:check && npm run lint:deps &&  npm run lint:server && npm run lint:ui",
     "lint:check": "./scripts/lint/check-filenames.sh && shellcheck scripts/*.sh scripts/*/*.sh deploy-templates/*.sh",
     "lint:deps": "deplint && depcheck --ignore-bin-package --ignores=@balena/jellyfish-*,shellcheck",
     "lint:catch": "catch-uncommitted --skip-node-versionbot-changes --exclude=VERSION,.versionbot/",
-    "lint:livechat": "(cd apps/livechat && npm run lint)",
     "lint:server": "(cd apps/server && npm run lint)",
     "lint:ui": "(cd apps/ui && npm run lint)",
     "test:unit": "npm run test:unit:ui && npm run test:unit:server",
     "test:unit:ui": "(cd apps/ui && npm run test:unit)",
     "test:unit:server": "(cd apps/server && npm run test:unit)",
     "test:e2e:sdk": "ava ./test/e2e/sdk/*.spec.js",
-    "test:e2e:livechat": "playwright test test/e2e/livechat",
     "test:e2e:ui": "playwright test test/e2e/ui",
     "test:e2e:server": "ava ./test/e2e/server/**/*.spec.js",
     "test:integration:server": "(cd apps/server && npm run test:integration)",
-    "docs": "(cd apps/livechat && npm run doc); (cd apps/server && npm run doc); (cd apps/ui && npm run doc)",
+    "docs": "(cd apps/server && npm run doc); (cd apps/ui && npm run doc)",
     "docs:diagram": "nomnoml docs/diagrams/architecture.noml > docs/assets/architecture.svg",
     "docs:summary": "./scripts/architecture-summary.sh > ARCHITECTURE.md",
     "clean": "rimraf apps/*/packages/*",
-    "postinstall": "(cd apps/livechat && npm i); (cd apps/server && npm i); (cd apps/ui && npm i)",
+    "postinstall": "(cd apps/server && npm i); (cd apps/ui && npm i)",
     "compose:build": "docker-compose build",
     "compose:up": "docker-compose up",
     "compose:test": "docker-compose -f docker-compose.yml -f docker-compose.test.yml up",
@@ -40,10 +38,8 @@
     "compose:database": "docker-compose up postgres redis haproxy",
     "push": "./scripts/push.sh",
     "push:lib": "./scripts/push-lib.js",
-    "build:livechat": "(cd apps/livechat && npm run build)",
     "build:server": "(cd apps/server && npm run build)",
     "build:ui": "(cd apps/ui && npm run build)",
-    "dev:livechat": "(cd apps/livechat && npm run dev)",
     "dev:server": "(cd apps/server && npm run dev)",
     "dev:ui": "(cd apps/ui && npm run dev)"
   },
